pkcs11-tool.c: Avoid warning:
pkcs11-tool.c:2000: warning: comparison between signed and unsigned


git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@5363 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/tools/pkcs11-tool.c
@@ -1997,7 +1997,7 @@ static void list_objects(CK_SESSION_HANDLE sess, CK_OBJECT_CLASS  object_class)
 			p11_fatal("C_FindObjects", rv);
 		if (count == 0)
 			break;
-		if (object_class == -1 || object_class == getCLASS(sess, object))
+		if ((int) object_class == -1 || object_class == getCLASS(sess, object))
 			show_object(sess, object);
 	}
 	p11->C_FindObjectsFinal(sess);
